+++ CHANGELOG.md
@@ -7,7 +7,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
-- managed object storage: add `TerminationProtection` field to `upcloud.ManagedObjectStorage` and `request.ModifyManagedObjectStorageRequest`
+- managed object storage: add `TerminationProtection` field to `upcloud.ManagedObjectStorage`, `request.CreateManagedObjectStorageRequest`, `request.ModifyManagedObjectStorageRequest` and `request.ReplaceManagedObjectStorageRequest`
 
 ## [8.40.0]